refactor(projects): split project assistant service
