+++ internal/projectassistant/service.go
@@ -1,7 +1,6 @@
 package projectassistant
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
@@ -286,985 +285,4 @@ func (s *Service) draftDeterministic(ctx context.Context, input DraftInput, draf
 		}
 	}
 	return s.Propose(ctx, proposalInput)
-}
-
-func (s *Service) Apply(ctx context.Context, proposal Proposal, confirmed bool) (ApplyResult, error) {
-	if s == nil {
-		return ApplyResult{}, ErrStoreNotConfigured
-	}
-	proposal.ID = strings.TrimSpace(proposal.ID)
-	if proposal.ID == "" {
-		return ApplyResult{}, fmt.Errorf("%w: proposal id is required", ErrInvalid)
-	}
-	if proposal.RequiresConfirmation && !confirmed {
-		return ApplyResult{}, ErrConfirmationRequired
-	}
-	for _, action := range proposal.Actions {
-		if err := validateActionShape(action); err != nil {
-			return ApplyResult{}, err
-		}
-	}
-	fingerprint, err := actionSetFingerprint(proposal.Actions)
-	if err != nil {
-		return ApplyResult{}, err
-	}
-
-	// Hold the apply lock through the mutation sequence so a proposal ID cannot
-	// race itself into duplicate durable writes. Progress is intentionally
-	// in-process for v0; retrying the exact same action set resumes after the
-	// last committed action instead of replaying earlier mutations.
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	progress, ok := s.applyProgress[proposal.ID]
-	if !ok {
-		progress = &applyProgress{fingerprint: fingerprint}
-		s.applyProgress[proposal.ID] = progress
-	}
-	if progress.complete {
-		return ApplyResult{}, fmt.Errorf("%w: proposal %q was already applied", ErrConflict, proposal.ID)
-	}
-	if progress.fingerprint != fingerprint {
-		return ApplyResult{}, fmt.Errorf("%w: proposal %q action set changed after partial apply", ErrConflict, proposal.ID)
-	}
-
-	for idx := len(progress.results); idx < len(proposal.Actions); idx++ {
-		action := proposal.Actions[idx]
-		result, err := s.applyAction(ctx, action)
-		if err != nil {
-			partial := ApplyResult{
-				ProposalID: proposal.ID,
-				Applied:    false,
-				Actions:    cloneActionResults(progress.results),
-			}
-			return partial, &ApplyError{
-				ProposalID:        proposal.ID,
-				FailedActionIndex: idx,
-				Result:            partial,
-				Err:               err,
-			}
-		}
-		progress.results = append(progress.results, result)
-	}
-
-	progress.complete = true
-
-	return ApplyResult{ProposalID: proposal.ID, Applied: true, Actions: cloneActionResults(progress.results)}, nil
-}
-
-func (s *Service) applyAction(ctx context.Context, action Action) (ActionResult, error) {
-	switch normalizeKind(action.Kind) {
-	case ActionCreateProject:
-		return s.applyCreateProject(ctx, action)
-	case ActionUpdateProject:
-		return s.applyUpdateProject(ctx, action)
-	case ActionAttachProjectRoot:
-		return s.applyAttachProjectRoot(ctx, action)
-	case ActionRemoveProjectRoot:
-		return s.applyRemoveProjectRoot(ctx, action)
-	case ActionSetProjectDefaults:
-		return s.applySetProjectDefaults(ctx, action)
-	case ActionMoveChatSession:
-		return s.applyMoveChatSession(ctx, action)
-	case ActionCreateRole:
-		return s.applyCreateRole(ctx, action)
-	case ActionCreateWorkItem:
-		return s.applyCreateWorkItem(ctx, action)
-	case ActionUpdateWorkItem:
-		return s.applyUpdateWorkItem(ctx, action)
-	case ActionCreateAssignment:
-		return s.applyCreateAssignment(ctx, action)
-	case ActionCreateHandoff:
-		return s.applyCreateHandoff(ctx, action)
-	case ActionCreateMemoryCandidate:
-		return s.applyCreateMemoryCandidate(ctx, action)
-	default:
-		return ActionResult{}, fmt.Errorf("%w: %s", ErrUnknownActionKind, action.Kind)
-	}
-}
-
-func validateActionShape(action Action) error {
-	kind := normalizeKind(action.Kind)
-	switch kind {
-	case ActionCreateProject, ActionUpdateProject, ActionAttachProjectRoot, ActionRemoveProjectRoot,
-		ActionSetProjectDefaults, ActionMoveChatSession, ActionCreateRole, ActionCreateWorkItem, ActionUpdateWorkItem,
-		ActionCreateAssignment, ActionCreateHandoff, ActionCreateMemoryCandidate:
-		if len(action.Patch) == 0 && kind != ActionRemoveProjectRoot {
-			return fmt.Errorf("%w: action %q patch is required", ErrInvalid, kind)
-		}
-		if kind == ActionCreateAssignment {
-			hasRuntimeLinks, err := assignmentPatchHasRuntimeLinks(action.Patch)
-			if err != nil {
-				return err
-			}
-			if hasRuntimeLinks {
-				return fmt.Errorf("%w: create_assignment proposals cannot bind chats, tasks, runs, messages, or snapshots", ErrInvalid)
-			}
-			var patch assignmentPatch
-			if err := decodePatch(action, &patch); err != nil {
-				return err
-			}
-			if err := validateAssignmentProposalBoundary(patch); err != nil {
-				return err
-			}
-		}
-		return nil
-	default:
-		return fmt.Errorf("%w: %s", ErrUnknownActionKind, action.Kind)
-	}
-}
-
-func validateAssignmentProposalBoundary(patch assignmentPatch) error {
-	if len(patch.ExecutionRef) > 0 {
-		return fmt.Errorf("%w: create_assignment proposals cannot bind chats, tasks, runs, messages, or snapshots", ErrInvalid)
-	}
-	if patch.Status != "" && patch.Status != projectwork.AssignmentStatusQueued {
-		return fmt.Errorf("%w: create_assignment proposals must create queued assignments", ErrInvalid)
-	}
-	return nil
-}
-
-func assignmentPatchHasRuntimeLinks(raw json.RawMessage) (bool, error) {
-	if len(raw) == 0 {
-		return false, nil
-	}
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(raw, &fields); err != nil {
-		return false, fmt.Errorf("%w: decode assignment patch keys: %v", ErrInvalid, err)
-	}
-	for _, key := range []string{"task_id", "run_id", "chat_session_id", "message_id", "context_snapshot_id", "execution_ref"} {
-		if _, ok := fields[key]; ok {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func (s *Service) applyCreateProject(ctx context.Context, action Action) (ActionResult, error) {
-	if s.projects == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch projectPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("proj")
-	} else if _, ok, err := s.projects.Get(ctx, id); err != nil {
-		return ActionResult{}, err
-	} else if ok {
-		return ActionResult{}, fmt.Errorf("%w: project %q already exists", ErrConflict, id)
-	}
-	roots, err := rootsForProjectPatch(patch, s.idgen)
-	if err != nil {
-		return ActionResult{}, err
-	}
-	project := projects.Project{
-		ID:                       id,
-		Name:                     patch.Name,
-		Description:              patch.Description,
-		Roots:                    roots,
-		DefaultProvider:          patch.DefaultProvider,
-		DefaultModel:             patch.DefaultModel,
-		DefaultAgentProfile:      patch.DefaultAgentProfile,
-		DefaultToolsEnabled:      patch.DefaultToolsEnabled,
-		DefaultWorkspaceMode:     patch.DefaultWorkspaceMode,
-		DefaultSystemPrompt:      patch.DefaultSystemPrompt,
-		DefaultCompactToolOutput: patch.DefaultCompactToolOutput,
-	}
-	created, err := s.projects.Create(ctx, project)
-	if err != nil {
-		return ActionResult{}, mapProjectErr(err)
-	}
-	return ActionResult{Kind: ActionCreateProject, ID: created.ID, Data: map[string]string{"project_id": created.ID}}, nil
-}
-
-func (s *Service) applyUpdateProject(ctx context.Context, action Action) (ActionResult, error) {
-	projectID := targetValue(action, "project_id")
-	if projectID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.project_id is required", ErrInvalid)
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return ActionResult{}, err
-	}
-	var patch updateProjectPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	updated, err := s.projects.Update(ctx, projectID, func(project *projects.Project) {
-		if patch.Name != nil {
-			project.Name = *patch.Name
-		}
-		if patch.Description != nil {
-			project.Description = *patch.Description
-		}
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectErr(err)
-	}
-	return ActionResult{Kind: ActionUpdateProject, ID: updated.ID, Data: map[string]string{"project_id": updated.ID}}, nil
-}
-
-func (s *Service) applyAttachProjectRoot(ctx context.Context, action Action) (ActionResult, error) {
-	projectID := targetValue(action, "project_id")
-	if projectID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.project_id is required", ErrInvalid)
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return ActionResult{}, err
-	}
-	var patch rootPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	root := rootFromPatch(patch, s.idgen)
-	updated, err := s.projects.Update(ctx, projectID, func(project *projects.Project) {
-		project.Roots = append(project.Roots, root)
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectErr(err)
-	}
-	return ActionResult{Kind: ActionAttachProjectRoot, ID: root.ID, Data: map[string]string{"project_id": updated.ID, "root_id": root.ID}}, nil
-}
-
-func (s *Service) applyRemoveProjectRoot(ctx context.Context, action Action) (ActionResult, error) {
-	projectID := targetValue(action, "project_id")
-	rootID := targetValue(action, "root_id")
-	if projectID == "" || rootID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.project_id and target.root_id are required", ErrInvalid)
-	}
-	project, err := s.requireProject(ctx, projectID)
-	if err != nil {
-		return ActionResult{}, err
-	}
-	if !projectHasRoot(project, rootID) {
-		return ActionResult{}, fmt.Errorf("%w: root %q", ErrNotFound, rootID)
-	}
-	updated, err := s.projects.Update(ctx, projectID, func(project *projects.Project) {
-		roots := project.Roots[:0]
-		for _, root := range project.Roots {
-			if root.ID != rootID {
-				roots = append(roots, root)
-			}
-		}
-		project.Roots = roots
-		if project.DefaultRootID == rootID {
-			project.DefaultRootID = ""
-		}
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectErr(err)
-	}
-	return ActionResult{Kind: ActionRemoveProjectRoot, ID: rootID, Data: map[string]string{"project_id": updated.ID, "root_id": rootID}}, nil
-}
-
-func (s *Service) applySetProjectDefaults(ctx context.Context, action Action) (ActionResult, error) {
-	projectID := targetValue(action, "project_id")
-	if projectID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.project_id is required", ErrInvalid)
-	}
-	project, err := s.requireProject(ctx, projectID)
-	if err != nil {
-		return ActionResult{}, err
-	}
-	var patch defaultsPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	if patch.DefaultRootID != nil && *patch.DefaultRootID != "" && !projectHasRoot(project, *patch.DefaultRootID) {
-		return ActionResult{}, fmt.Errorf("%w: root %q", ErrNotFound, *patch.DefaultRootID)
-	}
-	updated, err := s.projects.Update(ctx, projectID, func(project *projects.Project) {
-		if patch.DefaultRootID != nil {
-			project.DefaultRootID = *patch.DefaultRootID
-		}
-		if patch.DefaultProvider != nil {
-			project.DefaultProvider = *patch.DefaultProvider
-		}
-		if patch.DefaultModel != nil {
-			project.DefaultModel = *patch.DefaultModel
-		}
-		if patch.DefaultAgentProfile != nil {
-			project.DefaultAgentProfile = *patch.DefaultAgentProfile
-		}
-		if patch.DefaultToolsEnabled != nil {
-			project.DefaultToolsEnabled = patch.DefaultToolsEnabled
-		}
-		if patch.DefaultWorkspaceMode != nil {
-			project.DefaultWorkspaceMode = *patch.DefaultWorkspaceMode
-		}
-		if patch.DefaultSystemPrompt != nil {
-			project.DefaultSystemPrompt = *patch.DefaultSystemPrompt
-		}
-		if patch.DefaultCompactToolOutput != nil {
-			project.DefaultCompactToolOutput = patch.DefaultCompactToolOutput
-		}
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectErr(err)
-	}
-	return ActionResult{Kind: ActionSetProjectDefaults, ID: updated.ID, Data: map[string]string{"project_id": updated.ID}}, nil
-}
-
-func (s *Service) applyMoveChatSession(ctx context.Context, action Action) (ActionResult, error) {
-	if s.chats == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	sessionID := targetValue(action, "chat_session_id")
-	if sessionID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.chat_session_id is required", ErrInvalid)
-	}
-	var patch moveChatSessionPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := strings.TrimSpace(patch.ProjectID)
-	if projectID != "" {
-		if _, err := s.requireProject(ctx, projectID); err != nil {
-			return ActionResult{}, err
-		}
-	}
-	if _, ok, err := s.chats.Get(ctx, sessionID); err != nil {
-		return ActionResult{}, err
-	} else if !ok {
-		return ActionResult{}, fmt.Errorf("%w: chat session %q", ErrNotFound, sessionID)
-	}
-	updated, err := s.chats.UpdateSession(ctx, sessionID, func(session *chat.Session) {
-		session.ProjectID = projectID
-	})
-	if err != nil {
-		return ActionResult{}, err
-	}
-	return ActionResult{Kind: ActionMoveChatSession, ID: updated.ID, Data: map[string]string{"chat_session_id": updated.ID, "project_id": updated.ProjectID}}, nil
-}
-
-func (s *Service) applyCreateRole(ctx context.Context, action Action) (ActionResult, error) {
-	if s.work == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch rolePatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := patch.ProjectID
-	if projectID == "" {
-		projectID = targetValue(action, "project_id")
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("role")
-	}
-	role, err := s.work.CreateRole(ctx, projectwork.AgentRoleProfile{
-		ID:                  id,
-		ProjectID:           projectID,
-		Name:                patch.Name,
-		Description:         patch.Description,
-		Instructions:        patch.Instructions,
-		DefaultDriverKind:   patch.DefaultDriverKind,
-		DefaultProvider:     patch.DefaultProvider,
-		DefaultModel:        patch.DefaultModel,
-		DefaultAgentProfile: patch.DefaultAgentProfile,
-		SkillIDs:            append([]string(nil), patch.SkillIDs...),
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectWorkErr(err)
-	}
-	return ActionResult{Kind: ActionCreateRole, ID: role.ID, Data: map[string]string{"project_id": role.ProjectID, "role_id": role.ID}}, nil
-}
-
-func (s *Service) applyCreateWorkItem(ctx context.Context, action Action) (ActionResult, error) {
-	if s.work == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch workItemPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := patch.ProjectID
-	if projectID == "" {
-		projectID = targetValue(action, "project_id")
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("work")
-	}
-	item, err := s.work.CreateWorkItem(ctx, projectwork.WorkItem{
-		ID:              id,
-		ProjectID:       projectID,
-		Title:           patch.Title,
-		Brief:           patch.Brief,
-		Status:          patch.Status,
-		Priority:        patch.Priority,
-		OwnerRoleID:     patch.OwnerRoleID,
-		ReviewerRoleIDs: append([]string(nil), patch.ReviewerRoleIDs...),
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectWorkErr(err)
-	}
-	return ActionResult{Kind: ActionCreateWorkItem, ID: item.ID, Data: map[string]string{"project_id": item.ProjectID, "work_item_id": item.ID}}, nil
-}
-
-func (s *Service) applyUpdateWorkItem(ctx context.Context, action Action) (ActionResult, error) {
-	if s.work == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	projectID := targetValue(action, "project_id")
-	workItemID := targetValue(action, "work_item_id")
-	if projectID == "" || workItemID == "" {
-		return ActionResult{}, fmt.Errorf("%w: target.project_id and target.work_item_id are required", ErrInvalid)
-	}
-	if _, err := s.requireWorkItem(ctx, projectID, workItemID); err != nil {
-		return ActionResult{}, err
-	}
-	var patch updateWorkItemPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	item, err := s.work.UpdateWorkItem(ctx, projectID, workItemID, func(item *projectwork.WorkItem) {
-		if patch.Title != nil {
-			item.Title = *patch.Title
-		}
-		if patch.Brief != nil {
-			item.Brief = *patch.Brief
-		}
-		if patch.Status != nil {
-			item.Status = *patch.Status
-		}
-		if patch.Priority != nil {
-			item.Priority = *patch.Priority
-		}
-		if patch.OwnerRoleID != nil {
-			item.OwnerRoleID = *patch.OwnerRoleID
-		}
-		if patch.ReviewerRoleIDs != nil {
-			item.ReviewerRoleIDs = append([]string(nil), patch.ReviewerRoleIDs...)
-		}
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectWorkErr(err)
-	}
-	return ActionResult{Kind: ActionUpdateWorkItem, ID: item.ID, Data: map[string]string{"project_id": item.ProjectID, "work_item_id": item.ID}}, nil
-}
-
-func (s *Service) applyCreateAssignment(ctx context.Context, action Action) (ActionResult, error) {
-	if s.work == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch assignmentPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := patch.ProjectID
-	if projectID == "" {
-		projectID = targetValue(action, "project_id")
-	}
-	if _, err := s.requireWorkItem(ctx, projectID, patch.WorkItemID); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("asgn")
-	}
-	assignment, err := s.work.CreateAssignment(ctx, projectwork.Assignment{
-		ID:         id,
-		ProjectID:  projectID,
-		WorkItemID: patch.WorkItemID,
-		RoleID:     patch.RoleID,
-		DriverKind: patch.DriverKind,
-		Status:     patch.Status,
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectWorkErr(err)
-	}
-	return ActionResult{Kind: ActionCreateAssignment, ID: assignment.ID, Data: map[string]string{"project_id": assignment.ProjectID, "assignment_id": assignment.ID}}, nil
-}
-
-func (s *Service) applyCreateHandoff(ctx context.Context, action Action) (ActionResult, error) {
-	if s.work == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch handoffPatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := patch.ProjectID
-	if projectID == "" {
-		projectID = targetValue(action, "project_id")
-	}
-	if _, err := s.requireWorkItem(ctx, projectID, patch.WorkItemID); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("handoff")
-	}
-	handoff, err := s.work.CreateHandoff(ctx, projectwork.Handoff{
-		ID:                    id,
-		ProjectID:             projectID,
-		WorkItemID:            patch.WorkItemID,
-		SourceAssignmentID:    patch.SourceAssignmentID,
-		SourceRunID:           patch.SourceRunID,
-		SourceChatSessionID:   patch.SourceChatSessionID,
-		SourceMessageID:       patch.SourceMessageID,
-		TargetRoleID:          patch.TargetRoleID,
-		TargetAssignmentID:    patch.TargetAssignmentID,
-		TargetWorkItemID:      patch.TargetWorkItemID,
-		Title:                 patch.Title,
-		Summary:               patch.Summary,
-		RecommendedNextAction: patch.RecommendedNextAction,
-		LinkedArtifactIDs:     append([]string(nil), patch.LinkedArtifactIDs...),
-		LinkedMemoryIDs:       append([]string(nil), patch.LinkedMemoryIDs...),
-		ContextRefs:           append([]string(nil), patch.ContextRefs...),
-		Status:                patch.Status,
-		ProvenanceKind:        patch.ProvenanceKind,
-		TrustLabel:            patch.TrustLabel,
-		CreatedByRoleID:       patch.CreatedByRoleID,
-	})
-	if err != nil {
-		return ActionResult{}, mapProjectWorkErr(err)
-	}
-	return ActionResult{Kind: ActionCreateHandoff, ID: handoff.ID, Data: map[string]string{"project_id": handoff.ProjectID, "handoff_id": handoff.ID}}, nil
-}
-
-func (s *Service) applyCreateMemoryCandidate(ctx context.Context, action Action) (ActionResult, error) {
-	if s.memoryCandidates == nil {
-		return ActionResult{}, ErrStoreNotConfigured
-	}
-	var patch memoryCandidatePatch
-	if err := decodePatch(action, &patch); err != nil {
-		return ActionResult{}, err
-	}
-	projectID := patch.ProjectID
-	if projectID == "" {
-		projectID = targetValue(action, "project_id")
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return ActionResult{}, err
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = s.idgen("memcand")
-	}
-	candidate, err := s.memoryCandidates.CreateCandidate(ctx, memory.Candidate{
-		ID:                  id,
-		ProjectID:           projectID,
-		Title:               patch.Title,
-		Body:                patch.Body,
-		SuggestedKind:       patch.SuggestedKind,
-		SuggestedTrustLabel: patch.SuggestedTrustLabel,
-		SuggestedSourceKind: patch.SuggestedSourceKind,
-		SuggestedSourceID:   patch.SuggestedSourceID,
-		SourceRefs:          append([]memory.CandidateSourceRef(nil), patch.SourceRefs...),
-	})
-	if err != nil {
-		return ActionResult{}, mapMemoryErr(err)
-	}
-	return ActionResult{Kind: ActionCreateMemoryCandidate, ID: candidate.ID, Data: map[string]string{"project_id": candidate.ProjectID, "candidate_id": candidate.ID}}, nil
-}
-
-func (s *Service) requireProject(ctx context.Context, projectID string) (projects.Project, error) {
-	if s.projects == nil {
-		return projects.Project{}, ErrStoreNotConfigured
-	}
-	projectID = strings.TrimSpace(projectID)
-	if projectID == "" {
-		return projects.Project{}, fmt.Errorf("%w: project_id is required", ErrInvalid)
-	}
-	project, ok, err := s.projects.Get(ctx, projectID)
-	if err != nil {
-		return projects.Project{}, err
-	}
-	if !ok {
-		return projects.Project{}, fmt.Errorf("%w: project %q", ErrNotFound, projectID)
-	}
-	return project, nil
-}
-
-func (s *Service) requireWorkItem(ctx context.Context, projectID, workItemID string) (projectwork.WorkItem, error) {
-	if s.work == nil {
-		return projectwork.WorkItem{}, ErrStoreNotConfigured
-	}
-	if _, err := s.requireProject(ctx, projectID); err != nil {
-		return projectwork.WorkItem{}, err
-	}
-	workItemID = strings.TrimSpace(workItemID)
-	if workItemID == "" {
-		return projectwork.WorkItem{}, fmt.Errorf("%w: work_item_id is required", ErrInvalid)
-	}
-	item, ok, err := s.work.GetWorkItem(ctx, projectID, workItemID)
-	if err != nil {
-		return projectwork.WorkItem{}, err
-	}
-	if !ok {
-		return projectwork.WorkItem{}, fmt.Errorf("%w: work item %q", ErrNotFound, workItemID)
-	}
-	return item, nil
-}
-
-type projectPatch struct {
-	ID                       string      `json:"id,omitempty"`
-	Name                     string      `json:"name,omitempty"`
-	Description              string      `json:"description,omitempty"`
-	WorkspacePath            string      `json:"workspace_path,omitempty"`
-	WorkspaceKind            string      `json:"workspace_kind,omitempty"`
-	Roots                    []rootPatch `json:"roots,omitempty"`
-	DefaultProvider          string      `json:"default_provider,omitempty"`
-	DefaultModel             string      `json:"default_model,omitempty"`
-	DefaultAgentProfile      string      `json:"default_agent_profile,omitempty"`
-	DefaultToolsEnabled      *bool       `json:"default_tools_enabled,omitempty"`
-	DefaultWorkspaceMode     string      `json:"default_workspace_mode,omitempty"`
-	DefaultSystemPrompt      string      `json:"default_system_prompt,omitempty"`
-	DefaultCompactToolOutput *bool       `json:"default_compact_tool_output,omitempty"`
-}
-
-type updateProjectPatch struct {
-	Name        *string `json:"name,omitempty"`
-	Description *string `json:"description,omitempty"`
-}
-
-type rootPatch struct {
-	ID        string `json:"id,omitempty"`
-	Path      string `json:"path,omitempty"`
-	Kind      string `json:"kind,omitempty"`
-	GitRemote string `json:"git_remote,omitempty"`
-	GitBranch string `json:"git_branch,omitempty"`
-	Active    *bool  `json:"active,omitempty"`
-}
-
-type defaultsPatch struct {
-	DefaultRootID            *string `json:"default_root_id,omitempty"`
-	DefaultProvider          *string `json:"default_provider,omitempty"`
-	DefaultModel             *string `json:"default_model,omitempty"`
-	DefaultAgentProfile      *string `json:"default_agent_profile,omitempty"`
-	DefaultToolsEnabled      *bool   `json:"default_tools_enabled,omitempty"`
-	DefaultWorkspaceMode     *string `json:"default_workspace_mode,omitempty"`
-	DefaultSystemPrompt      *string `json:"default_system_prompt,omitempty"`
-	DefaultCompactToolOutput *bool   `json:"default_compact_tool_output,omitempty"`
-}
-
-type moveChatSessionPatch struct {
-	ProjectID string `json:"project_id"`
-}
-
-type rolePatch struct {
-	ID                  string   `json:"id,omitempty"`
-	ProjectID           string   `json:"project_id,omitempty"`
-	Name                string   `json:"name,omitempty"`
-	Description         string   `json:"description,omitempty"`
-	Instructions        string   `json:"instructions,omitempty"`
-	DefaultDriverKind   string   `json:"default_driver_kind,omitempty"`
-	DefaultProvider     string   `json:"default_provider,omitempty"`
-	DefaultModel        string   `json:"default_model,omitempty"`
-	DefaultAgentProfile string   `json:"default_agent_profile,omitempty"`
-	SkillIDs            []string `json:"skill_ids,omitempty"`
-}
-
-type workItemPatch struct {
-	ID              string   `json:"id,omitempty"`
-	ProjectID       string   `json:"project_id,omitempty"`
-	Title           string   `json:"title,omitempty"`
-	Brief           string   `json:"brief,omitempty"`
-	Status          string   `json:"status,omitempty"`
-	Priority        string   `json:"priority,omitempty"`
-	OwnerRoleID     string   `json:"owner_role_id,omitempty"`
-	ReviewerRoleIDs []string `json:"reviewer_role_ids,omitempty"`
-}
-
-type updateWorkItemPatch struct {
-	Title           *string  `json:"title,omitempty"`
-	Brief           *string  `json:"brief,omitempty"`
-	Status          *string  `json:"status,omitempty"`
-	Priority        *string  `json:"priority,omitempty"`
-	OwnerRoleID     *string  `json:"owner_role_id,omitempty"`
-	ReviewerRoleIDs []string `json:"reviewer_role_ids,omitempty"`
-}
-
-type assignmentPatch struct {
-	ID           string          `json:"id,omitempty"`
-	ProjectID    string          `json:"project_id,omitempty"`
-	WorkItemID   string          `json:"work_item_id,omitempty"`
-	RoleID       string          `json:"role_id,omitempty"`
-	DriverKind   string          `json:"driver_kind,omitempty"`
-	Status       string          `json:"status,omitempty"`
-	ExecutionRef json.RawMessage `json:"execution_ref,omitempty"`
-}
-
-type handoffPatch struct {
-	ID                    string   `json:"id,omitempty"`
-	ProjectID             string   `json:"project_id,omitempty"`
-	WorkItemID            string   `json:"work_item_id,omitempty"`
-	SourceAssignmentID    string   `json:"source_assignment_id,omitempty"`
-	SourceRunID           string   `json:"source_run_id,omitempty"`
-	SourceChatSessionID   string   `json:"source_chat_session_id,omitempty"`
-	SourceMessageID       string   `json:"source_message_id,omitempty"`
-	TargetRoleID          string   `json:"target_role_id,omitempty"`
-	TargetAssignmentID    string   `json:"target_assignment_id,omitempty"`
-	TargetWorkItemID      string   `json:"target_work_item_id,omitempty"`
-	Title                 string   `json:"title,omitempty"`
-	Summary               string   `json:"summary,omitempty"`
-	RecommendedNextAction string   `json:"recommended_next_action,omitempty"`
-	LinkedArtifactIDs     []string `json:"linked_artifact_ids,omitempty"`
-	LinkedMemoryIDs       []string `json:"linked_memory_ids,omitempty"`
-	ContextRefs           []string `json:"context_refs,omitempty"`
-	Status                string   `json:"status,omitempty"`
-	ProvenanceKind        string   `json:"provenance_kind,omitempty"`
-	TrustLabel            string   `json:"trust_label,omitempty"`
-	CreatedByRoleID       string   `json:"created_by_role_id,omitempty"`
-}
-
-type memoryCandidatePatch struct {
-	ID                  string                      `json:"id,omitempty"`
-	ProjectID           string                      `json:"project_id,omitempty"`
-	Title               string                      `json:"title,omitempty"`
-	Body                string                      `json:"body,omitempty"`
-	SuggestedKind       string                      `json:"suggested_kind,omitempty"`
-	SuggestedTrustLabel string                      `json:"suggested_trust_label,omitempty"`
-	SuggestedSourceKind string                      `json:"suggested_source_kind,omitempty"`
-	SuggestedSourceID   string                      `json:"suggested_source_id,omitempty"`
-	SourceRefs          []memory.CandidateSourceRef `json:"source_refs,omitempty"`
-}
-
-func decodePatch(action Action, target any) error {
-	if len(action.Patch) == 0 {
-		return nil
-	}
-	if err := json.Unmarshal(action.Patch, target); err != nil {
-		return fmt.Errorf("%w: decode %s patch: %v", ErrInvalid, normalizeKind(action.Kind), err)
-	}
-	return nil
-}
-
-func rootFromPatch(patch rootPatch, idgen IDGenerator) projects.Root {
-	active := true
-	if patch.Active != nil {
-		active = *patch.Active
-	}
-	id := strings.TrimSpace(patch.ID)
-	if id == "" {
-		id = idgen("root")
-	}
-	return projects.Root{
-		ID:        id,
-		Path:      patch.Path,
-		Kind:      patch.Kind,
-		GitRemote: patch.GitRemote,
-		GitBranch: patch.GitBranch,
-		Active:    active,
-	}
-}
-
-func rootsFromPatch(patches []rootPatch, idgen IDGenerator) []projects.Root {
-	roots := make([]projects.Root, 0, len(patches))
-	for _, patch := range patches {
-		roots = append(roots, rootFromPatch(patch, idgen))
-	}
-	return roots
-}
-
-func rootsForProjectPatch(patch projectPatch, idgen IDGenerator) ([]projects.Root, error) {
-	roots := rootsFromPatch(patch.Roots, idgen)
-	workspacePath := strings.TrimSpace(patch.WorkspacePath)
-	workspaceKind := strings.TrimSpace(patch.WorkspaceKind)
-	if workspacePath != "" {
-		if len(roots) > 0 {
-			return nil, fmt.Errorf("%w: workspace_path cannot be combined with roots", ErrInvalid)
-		}
-		roots = []projects.Root{{
-			ID:     idgen("root"),
-			Path:   workspacePath,
-			Kind:   workspaceKind,
-			Active: true,
-		}}
-	} else if workspaceKind != "" {
-		return nil, fmt.Errorf("%w: workspace_kind requires workspace_path", ErrInvalid)
-	}
-	return roots, nil
-}
-
-func projectHasRoot(project projects.Project, rootID string) bool {
-	rootID = strings.TrimSpace(rootID)
-	for _, root := range project.Roots {
-		if root.ID == rootID {
-			return true
-		}
-	}
-	return false
-}
-
-func targetValue(action Action, key string) string {
-	if action.Target == nil {
-		return ""
-	}
-	return strings.TrimSpace(action.Target[key])
-}
-
-func normalizeKind(kind string) string {
-	return strings.TrimSpace(kind)
-}
-
-func cloneActions(actions []Action) []Action {
-	if actions == nil {
-		return nil
-	}
-	cloned := make([]Action, len(actions))
-	for idx, action := range actions {
-		cloned[idx] = Action{
-			Kind:   action.Kind,
-			Target: cloneStringMap(action.Target),
-			Patch:  append(json.RawMessage(nil), action.Patch...),
-			Reason: action.Reason,
-		}
-	}
-	return cloned
-}
-
-func actionSetFingerprint(actions []Action) (string, error) {
-	items := make([]actionFingerprint, 0, len(actions))
-	for _, action := range actions {
-		patch, err := compactPatch(action.Patch)
-		if err != nil {
-			return "", err
-		}
-		items = append(items, actionFingerprint{
-			Kind:   normalizeKind(action.Kind),
-			Target: cloneStringMap(action.Target),
-			Patch:  patch,
-			Reason: strings.TrimSpace(action.Reason),
-		})
-	}
-	encoded, err := json.Marshal(items)
-	if err != nil {
-		return "", fmt.Errorf("%w: encode action fingerprint: %v", ErrInvalid, err)
-	}
-	return string(encoded), nil
-}
-
-func compactPatch(raw json.RawMessage) (json.RawMessage, error) {
-	if len(raw) == 0 {
-		return nil, nil
-	}
-	var buf bytes.Buffer
-	if err := json.Compact(&buf, raw); err != nil {
-		return nil, fmt.Errorf("%w: invalid action patch json", ErrInvalid)
-	}
-	return append(json.RawMessage(nil), buf.Bytes()...), nil
-}
-
-func cloneActionResults(results []ActionResult) []ActionResult {
-	if results == nil {
-		return nil
-	}
-	cloned := make([]ActionResult, len(results))
-	for idx, result := range results {
-		cloned[idx] = ActionResult{
-			Kind: result.Kind,
-			ID:   result.ID,
-			Data: cloneStringMap(result.Data),
-		}
-	}
-	return cloned
-}
-
-func cloneStringMap(input map[string]string) map[string]string {
-	if input == nil {
-		return nil
-	}
-	out := make(map[string]string, len(input))
-	for key, value := range input {
-		out[key] = value
-	}
-	return out
-}
-
-type draftRequest struct {
-	title string
-	brief string
-}
-
-func draftRequestParts(request string) draftRequest {
-	lines := strings.Split(strings.ReplaceAll(strings.ReplaceAll(request, "\r\n", "\n"), "\r", "\n"), "\n")
-	parts := make([]string, 0, len(lines))
-	for _, line := range lines {
-		if trimmed := strings.TrimSpace(line); trimmed != "" {
-			parts = append(parts, trimmed)
-		}
-	}
-	if len(parts) == 0 {
-		return draftRequest{}
-	}
-	return draftRequest{title: parts[0], brief: strings.Join(parts[1:], "\n\n")}
-}
-
-func validDraftDriverKind(kind string) bool {
-	switch kind {
-	case projectwork.AssignmentDriverHecateTask, projectwork.AssignmentDriverExternalAgent:
-		return true
-	default:
-		return false
-	}
-}
-
-func mustRawJSON(value any) json.RawMessage {
-	encoded, err := json.Marshal(value)
-	if err != nil {
-		panic(err)
-	}
-	return encoded
-}
-
-func firstNonEmpty(values ...string) string {
-	for _, value := range values {
-		if trimmed := strings.TrimSpace(value); trimmed != "" {
-			return trimmed
-		}
-	}
-	return ""
-}
-
-func mapProjectErr(err error) error {
-	switch {
-	case errors.Is(err, projects.ErrNotFound):
-		return fmt.Errorf("%w: %v", ErrNotFound, err)
-	case errors.Is(err, projects.ErrInvalid):
-		return fmt.Errorf("%w: %v", ErrInvalid, err)
-	case errors.Is(err, projects.ErrAlreadyExists):
-		return fmt.Errorf("%w: %v", ErrConflict, err)
-	default:
-		return err
-	}
-}
-
-func mapProjectWorkErr(err error) error {
-	switch {
-	case errors.Is(err, projectwork.ErrNotFound):
-		return fmt.Errorf("%w: %v", ErrNotFound, err)
-	case errors.Is(err, projectwork.ErrInvalid):
-		return fmt.Errorf("%w: %v", ErrInvalid, err)
-	case errors.Is(err, projectwork.ErrDuplicate), errors.Is(err, projectwork.ErrDuplicateRole):
-		return fmt.Errorf("%w: %v", ErrConflict, err)
-	case errors.Is(err, projectwork.ErrBuiltInRole):
-		return fmt.Errorf("%w: %v", ErrInvalid, err)
-	default:
-		return err
-	}
-}
-
-func mapMemoryErr(err error) error {
-	switch {
-	case errors.Is(err, memory.ErrNotFound):
-		return fmt.Errorf("%w: %v", ErrNotFound, err)
-	case errors.Is(err, memory.ErrInvalid):
-		return fmt.Errorf("%w: %v", ErrInvalid, err)
-	case errors.Is(err, memory.ErrAlreadyExists), errors.Is(err, memory.ErrConflict):
-		return fmt.Errorf("%w: %v", ErrConflict, err)
-	default:
-		return err
-	}
 }
